Move GeoHash specific unit tests own file
